test: reward 결과 테스트 보강

+++ src/main/java/game/GameResult.java
@@ -17,7 +17,7 @@ public enum GameResult {
     }
 
     public double calculateReward(Participant participant, Dealer dealer) {
-        double reward = (double) (participant.getBetMoney() * this.rewardRate);
+        double reward = participant.getBetMoney() * this.rewardRate;
 
         dealer.setMoney(dealer.getBetMoney() - reward);
         return reward;
